requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.5
fastapi-cli==0.0.5
fonttools==4.55.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pydantic==2.10.1
pydantic_core==2.27.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.2
rich==13.9.4
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
threadpoolctl==3.5.0
tqdm==4.67.1
typer==0.13.1
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.1
watchfiles==1.0.0
websockets==14.1
